requirements.txt
Django
djangorestframework
django-filter
djangorestframework-simplejwt
psycopg2-binary
django-environ
coverage
drf-yasg
black